Clean some more shared_ptrs. Increase warning level to 4 + warning as error option. Fix them. Replace some string copies with std::string_view.

+++ SYEngine/Diagnostics/Logger.cpp
@@ -1,7 +1,5 @@
 
 
-#include <cstdio>
-#include <exception>
 
 #include "Logger.h"
 
@@ -76,7 +74,7 @@ void Logger::Log(bool showTimestamp, eLogType logType,const char* format, ...) c
 }
 
 
-Logger::Logger(DWORD PID):
+Logger::Logger(unsigned long PID):
   _PID(PID), 
   _hStdOut(GetStdHandle(STD_OUTPUT_HANDLE))
 {
